feat: extract ShowData, use on more pages
