+++ dropbox-hacking-photo-manager-client/src/components/ShowData.tsx
