+++ dropbox-hacking-photo-manager-client/src/pages/next-gen/contentHash/ShowContentHashResult.tsx
@@ -1,5 +1,6 @@
 import type { MediainfoFromHash } from "@blaahaj/dropbox-hacking-mediainfo-db";
 import GeoMap from "@components/map/GeoMap";
+import ShowData from "@components/ShowData";
 import { GPSLatLong } from "dropbox-hacking-photo-manager-shared";
 import type { ContentHashCollection } from "dropbox-hacking-photo-manager-shared/serverSideFeeds";
 import * as L from "leaflet";
@@ -7,7 +8,6 @@ import React from "react";
 
 import EditablePhotoEntry from "./EditablePhotoEntry";
 import ImagePreview from "./imagePreview";
-import ShowData from "./ShowData";
 import SummariseExif from "./SummariseExif";
 import SummariseMediaInfo from "./SummariseMediaInfo";
 import SummariseNamedFiles from "./SummariseNamedFiles";
